project/back-end/requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==3.2.16
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
filelock==3.10.7
platformdirs==3.2.0
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
virtualenv==20.21.0
